Added documentation, and some spline calculations

+++ src/PathPlanner.java
@@ -12,8 +12,8 @@ public class PathPlanner {
 	
 	//waypoints are entered as (x1, y1, t1, x2, y2, t2,...), units are arbitrary.
 	//TODO create a more dynamic system of entering waypoints
-	private static int[] wayPoints = {2,    2,    3,
-			                          3,    1,    5,};
+	private static float[] wayPoints = {2,    2,    3,
+			                            3,    1,    5,};
 	
 	public static void main(String[] args)
 	{
@@ -24,7 +24,6 @@ public class PathPlanner {
 	{
 		if(wayPoints.length % 3 == 0)
 			spline = new Spline(wayPoints);
-
 	}
 	
 }
